Fix dependency of ConductorStateTest on AggregatorHandler
