+++ arangod/Pregel/Conductor/ExecutionStates/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(arango_pregel_conductor PRIVATE
+target_sources(arango_pregel_with_actors PRIVATE
   CreateWorkersState.cpp
   LoadingState.cpp
   ComputingState.cpp)
